Refactored hooks
