+++ assets/js/hooks/collapsible_open.js
@@ -1,0 +1,7 @@
+const CollapsibleOpen = {
+  mounted() {
+    this.el.open = true;
+  },
+};
+
+export default CollapsibleOpen;
